firmware: Refactor output module to reduce interval between commands
In the previous design, command to left/right servo/motors
were sent 4ms apart allowing for a simpler design.
However, that would mean that the motor on the left would apply
a command before the motor on the right.

This new design partially fixes this issue. Commands to servos
are sent at the same time, followed 2.5ms later by motor
commands. It was not possible to fix the issue due to a lack
of timers (timer 3 & 4 are the only ones available).

Signed-off-by: Francois Berder <fberder@outlook.fr>